% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_pal.R
\name{colour_pal}
\alias{colour_pal}
\title{Get Colour Palettes}
\usage{
colour_pal(
  pal_name,
  n,
  assign,
  type = c("discrete", "discrete_as", "continuous")
)
}
\arguments{
\item{pal_name}{Name of desired palette or individual colour.}

\item{n}{Number of desired colours.}

\item{assign}{Vector of corresponding levels/cetgories}

\item{type}{Either "continuous", "discrete" (unnamed), or "discrete_as" (named/assigned). Default = "discrete".}
}
\value{
A vector of colours (discrete) or list of colours with names (discrete_as)
}
\description{
Generate colours palettes for graphs.
}
\examples{
# Return full palette
colour_pal("catExtended")

# Return individual colour
colour_pal("Jaffa")

# Return palette with 5 colours and assigned levels for each colour
colour_pal("divBlueGreen", 5, c("Very Likely","Likely","Neutral","Unlikely","Very Unlikely"))
}
