% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab.R
\name{crosstab}
\alias{crosstab}
\title{Produce Crosstabs / Contingency Tables for Analysis}
\usage{
crosstab(
  data,
  rowVar,
  colVar,
  weight = NULL,
  totals = TRUE,
  round_decimals = NULL,
  statistics = FALSE,
  plot = FALSE,
  format = c("df_long", "df_wide", "csv", "statistics"),
  convert_to = c("percent", "frequency"),
  yLab = "Population (\%)",
  adjustX = FALSE
)
}
\arguments{
\item{data}{A data frame containing survey data. This parameter is required.}

\item{rowVar}{The independent variable represented in rows (side of the table).}

\item{colVar}{The dependent variable represented in columns (top of the table).}

\item{weight}{An optional variable containing weight factors for the analysis.}

\item{totals}{Logical; if \code{TRUE}, includes a totals column in the output (default is \code{TRUE}).}

\item{round_decimals}{Optional; number of decimal points for rounding the data (default is \code{NULL}).}

\item{statistics}{Logical; if \code{TRUE}, calculates and prints Chi-Square, degrees of freedom (DF), Cramer's V,
and p-value (default is \code{FALSE}).}

\item{plot}{Logical; if \code{TRUE}, generates a plot for visual representation of the crosstab data (default is \code{FALSE}).}

\item{format}{Specifies the output format: 'df_long', 'df_wide', 'csv', or 'statistics' (default is 'df_long').}

\item{convert_to}{Determines conversion type: 'percent' or 'frequency' (default is 'percent').}

\item{yLab}{Title for the Y-axis, default is "Population (\%)".}

\item{adjustX}{Logical; if \code{TRUE}, adjusts the X-axis labels to a 45-degree angle for better readability
(default is \code{FALSE}).}
}
\value{
Depending on the chosen format, returns a data frame in either long or wide format,
containing row-wise percentages and column-wise totals. If \code{plot} is \code{TRUE}, a plot object is also generated.
}
\description{
The \code{crosstab} function generates crosstabs (contingency tables) from survey data,
allowing for a variety of formatting and analytical options. It can produce tables with
statistical measures, create visual plots, and handle weighted data.
}
\examples{
\dontrun{
  # Example: Create crosstabs with a long data frame output using weighted data
  crosstab_result <- crosstab(data = dataset,
                              rowVar = "Q1",
                              colVar = "Gender",
                              weight = "wgtvar",
                              totals = FALSE,
                              round_decimals = 2)
}
}
