% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get Dataset}
\usage{
get_data(...)
}
\arguments{
\item{...}{A string, naming the desired dataset.
Available datasets include:
\itemize{
\item \code{survey}: Wave 25 of the 2014-2023 British Election Study Internet Panel
}}
}
\value{
\code{get_data()} returns a data frame.
}
\description{
\code{get_data()} is a helper function for loading datasets. It provides the user with the ability to set
the name of the data frame within their own environment (e.g. df).
}
\examples{
df <- get_data("survey")
}
