% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file.R
\name{get_file}
\alias{get_file}
\title{Retrieve and Process File from Various Sources}
\usage{
get_file(file_path, source = c("local", "onedrive", "web"), row_no = 0)
}
\arguments{
\item{file_path}{The path, ID, or URL of the file to be retrieved.}

\item{source}{The source of the file: 'local', 'onedrive', or 'web'
(default is 'local').}

\item{row_no}{The number of rows to skip at the beginning of the file, applicable
for CSV files (default is 0).}
}
\value{
A data frame containing the contents of the file.
}
\description{
\code{get_file} handles the retrieval and initial processing of a file from different
sources, including local storage, OneDrive, and directly from the web.
It also preprocesses the file based on its type.
}
\details{
The function first identifies the file type based on its extension, then retrieves
the file from the specified source using \code{authenticate_source}. After retrieval,
the file is preprocessed according to its type using \code{preprocess_file_type},
which handles different formats like CSV and SAV.
}
\examples{
\dontrun{
  # Retrieve a local CSV file
  data <- get_file("path/to/local/file.csv")

  # Retrieve a file from OneDrive
  data <- get_file("file-id", source = "onedrive")

  # Retrieve and preprocess a Google Drive file, skipping the first row
  data <- get_file("file-id", source = "googledrive", row_no = 1)

  # Retrieve a file directly from a web URL
  data <- get_file("https://example.com/data.csv", source = "web")
}

}
