% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_question.R
\name{get_question}
\alias{get_question}
\title{Retrieve Original Question Label}
\usage{
get_question(data, var)
}
\arguments{
\item{data}{A data frame containing the variable for which the question label
is to be retrieved.}

\item{var}{The name of the variable within the data frame whose question label
is being sought. This should be a string corresponding to the column name in
\code{data}.}
}
\value{
Returns the original question label as a string. If no label is
associated with the variable, the variable name is returned.
}
\description{
This helper function retrieves the original question label associated with a
variable in a data frame. If no label attribute is present, it returns the
variable name itself.
}
\details{
The function checks for a \code{label} attribute on the specified variable in the
data frame. If the \code{label} attribute exists, it is returned as the question
label; otherwise, the name of the variable (\code{var}) is returned.
}
\note{
This function is used by \code{compile}, \code{crosstab}, and \code{plot_popn}.
}
