% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_vars.R
\name{grid_vars}
\alias{grid_vars}
\title{Binary Survey Data}
\usage{
grid_vars(data, vars, group, weight)
}
\arguments{
\item{data}{A data frame containing survey data. This parameter is required.}

\item{vars}{A list of column names and the renamed variable. This parameter is required.}

\item{group}{A variable overlay to compare between groups. This parameter is optional.}

\item{weight}{Variable containing weight factors. This variable is optional.}
}
\value{
a data frame of variables from a grid survey question.
}
\description{
Pivot binary grid survey questions longer.
}
\examples{
\dontrun{
# Create list
vars <- list(Q1a = "Art",
             Q1b = "Automobiles",
             Q1c = "Birdwatching")
grid_vars(dataset,
            vars = vars,
            group = "gender",
            weight = "wgtvar")

#      Question Response gender     Freq Perc
#           Art      Yes   Male 275.3617   46
#   Automobiles      Yes   Male 320.1372   53
#  Birdwatching      Yes   Male 310.4357   52
#           Art      Yes Female 204.7525   48
#   Automobiles      Yes Female 212.0209   49
#  Birdwatching      Yes Female 203.9380   47
# ...
}
}
