% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_mean.R
\name{grp_mean}
\alias{grp_mean}
\title{Calculate Grouped Mean in Survey Data}
\usage{
grp_mean(
  data,
  meanVar,
  groups,
  weight = NULL,
  set_names = NULL,
  round_decimals = NULL
)
}
\arguments{
\item{data}{A data frame containing survey data.}

\item{meanVar}{The variable for which the mean is to be calculated; must be numeric.}

\item{groups}{Variables used for grouping; one or many variables can be specified.}

\item{weight}{An optional variable containing weight factors for calculating the weighted mean.}

\item{set_names}{An optional vector of custom column names for the output data frame.}

\item{round_decimals}{An optional numeric value specifying the number of decimal
places for rounding the mean values.}
}
\value{
A data frame containing the calculated means for each group, with the option
to include weighted means and custom column names.
}
\description{
\code{grp_mean} calculates the mean (either unweighted or weighted) of a specified variable,
grouped by one or more variables in survey data.
}
\details{
The function can calculate either an unweighted mean if \code{weight} is \code{NULL}, or a
weighted mean if \code{weight} is provided.
If \code{weight} is provided, it calculates a weighted mean, accounting for the interaction
between group variables. The function is designed to work with group interactions
split by a full stop (.), which might affect the handling of group values containing periods.
The results can be rounded to a specified
number of decimal places, and custom column names can be set for the output data frame.
}
\examples{
\dontrun{
  # Calculate unweighted mean of 'age' grouped by 'gender'
  df <- grp_mean(dataset, meanVar = "age", groups = "gender")

  # Calculate weighted mean of 'age' grouped by 'gender', with custom column names
  df <- grp_mean(dataset, meanVar = "age", groups = "gender",
                 weight = "wgtvar", set_names = c("Gender", "Average Age"))
}

}
