% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_likert.R
\name{plot_likert}
\alias{plot_likert}
\title{Visualise Likert Scales with Stacked, Divergent, and Facetted Bar Plots}
\usage{
plot_likert(
  data,
  vars,
  varLevels = NULL,
  group = NULL,
  weight = NULL,
  type = c("stacked", "divergent", "facetted"),
  neutrals = c("no_change", "right", "exclude"),
  total = FALSE,
  NET = FALSE,
  title = NULL,
  subtitle = NULL,
  order_by = NULL,
  colours = colour_pal("divRedBlue"),
  legend = c("top", "right", "left", "bottom", "none"),
  nrow = 1,
  width = 0.8,
  ratio = 6
)
}
\arguments{
\item{data}{A data frame containing survey data.}

\item{vars}{A vector of variables containing Likert responses.}

\item{varLevels}{A vector or a named list with 'left', 'neutral', and 'right' elements specifying the levels of the Likert scale.}

\item{group}{An optional grouping variable for creating facetted plots. If specified, \code{vars} should contain only a single variable.}

\item{weight}{An optional weighting variable for the survey data.}

\item{type}{The type of Likert plot: 'stacked', 'divergent', or 'facetted'.}

\item{neutrals}{Handling of neutral responses in the plot. Can be 'no_change', 'right', or 'exclude'.}

\item{total}{If TRUE and \code{group} is specified, includes an option for the total population.}

\item{NET}{If TRUE, provides a Net Rating Score (positive minus negative responses).}

\item{title}{Title of the plot.}

\item{subtitle}{Subtitle of the plot.}

\item{order_by}{Specifies the ordering of responses in the plot.}

\item{colours}{Colour palette for the variables. Default = colour_pal("divRedBlue").}

\item{legend}{Position of the legend in the plot ('top', 'right', 'left', 'bottom', 'none').}

\item{nrow}{Number of rows for the legend, if applicable. Default = 1.}

\item{width}{Width of the bars in the plot. Default = 0.8.}

\item{ratio}{Aspect ratio of the plot. Default = 6.}
}
\value{
A \code{ggplot2} object representing the Likert plot.
}
\description{
\code{plot_likert} generates various types of Likert plots (stacked, divergent, or facetted) for visualising
survey data. It is versatile in handling different plot types, the inclusion of NET scores, and custom orderings.
}
\examples{
\dontrun{
  data <- survey_data
  vars <- c(Q1 = "Category 1", Q2 = "Category 2")
  varLevels <- list(left = c("Strongly Disagree", "Disagree"),
                    neutral = "Neutral",
                    right = c("Agree", "Strongly Agree"))
  plot_likert(data, vars, varLevels, type = "divergent")
}

}
