% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_parliament.R
\name{plot_parliament}
\alias{plot_parliament}
\title{Parliament Plot}
\usage{
plot_parliament(data, group)
}
\arguments{
\item{data}{A data frame containing survey data. This parameter is required.}

\item{group}{A variable overlay to compare between groups. This parameter is optional.}
}
\value{
ggplot2 plot of the parliamentary seats
}
\description{
Visualise the number of seats in parliament by party.
}
\examples{
\dontrun{
# Create plot using age groups and age intervals included to return average age
plot_parliament(dataset,)
}
}
