% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_popn.R
\name{plot_popn}
\alias{plot_popn}
\title{Visualise Population Profile of Survey Data}
\usage{
plot_popn(
  data,
  xVar,
  yVar,
  group = NULL,
  weight = NULL,
  meanVar = NULL,
  colours = NULL,
  title = "Population Structure",
  subtitle = NULL,
  xLab = "Population",
  yLab = "Age",
  addLabels = FALSE,
  thresholdLab = 3,
  nudgeLab = 0.2,
  sizeLab = 11,
  faceLab = c("plain", "bold", "italic", "bold.italic")
)
}
\arguments{
\item{data}{A data frame containing survey data.}

\item{xVar}{The gender variable (or a similar categorical variable) for the x-axis.}

\item{yVar}{The age group variable for the y-axis.}

\item{group}{An optional variable for additional grouping or comparison.}

\item{weight}{An optional weighting variable for the analysis.}

\item{meanVar}{An optional numeric variable to include average values in the plot.}

\item{colours}{A vector of three colors for male, female, and total. This parameter is optional.}

\item{title}{The title of the plot; defaults to "Population Structure".}

\item{subtitle}{The subtitle for the plot, particularly useful in grouped plots;
defaults to the question associated with the group variable.}

\item{xLab}{The title for the x-axis; defaults to "Population (\%)".}

\item{yLab}{The title for the y-axis; defaults to "Age".}

\item{addLabels}{A logical indicating whether to add percentage labels to the plot; defaults to \code{FALSE}.}

\item{thresholdLab}{A numeric threshold for label placement inside or outside the bars.}

\item{nudgeLab}{A numeric value to adjust the horizontal position of labels.}

\item{sizeLab}{The font size for labels.}

\item{faceLab}{The font style for labels; can be "plain", "bold", "italic", or "bold.italic".}
}
\value{
A \code{ggplot2} plot representing the population profile based on the provided survey data.
}
\description{
\code{plot_popn} creates a visual representation of the population profile using survey data.
It can illustrate the distribution of age groups across genders or other categories,
and optionally includes average age or other statistics.
}
\examples{
\dontrun{
  # Example: Create a population plot with age and gender
  plot_popn(dataset,
            xVar = "gender",
            yVar = "age_categories",
            weight = "wgtvar",
            meanVar = "age")
}

}
