% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survey_wt}
\alias{survey_wt}
\title{Wave 25 of the 2014-2023 British Election Study Internet Panel}
\format{
A data frame with 30,407 rows and 6 variables:
\describe{
\item{wt}{weight variable}
\item{age}{Age}
\item{ageGroup}{Age group}
\item{gender}{Gender}
\item{p_past_vote_2019}{2019 GE vote choice}
\item{p_turnout_2019}{2019 GE turnout}
}
}
\source{
This dataset was sourced from the British Electoral Study website (Wave 25).
The raw data are available from \url{https://www.britishelectionstudy.com/data-object/wave-25-of-the-2014-2023-british-election-study-internet-panel/}.
}
\usage{
survey_wt
}
\description{
The \code{survey_wt} dataset is from the British Electoral Study. It is purely as a way to practice
using the weighting functions within this and other packages.
}
\details{
A full description of themethodology and questionnaire can be found here:
\url{https://www.britishelectionstudy.com/wp-content/uploads/2023/11/Bes_wave25Documentation.pdf}
}
\keyword{datasets}
