% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_scg.R
\name{theme_scg}
\alias{theme_scg}
\title{Theme for scg Plots}
\usage{
theme_scg(base_size = 10, base_font = "")
}
\arguments{
\item{base_size}{Size of font.}

\item{base_font}{Font family. See \code{names(pdfFonts())} for options}
}
\value{
A ggplot2 theme
}
\description{
A theme for scg plots.
}
\examples{
\dontrun{
ggplot(data=df, aes(x = x, y= y, fill=reorder(group, y))) +
  geom_bar(stat="identity", width=0.8, position = position_dodge(width=0.9), alpha=1) +
  scale_fill_manual(values=colour_pal("catExtended")) +
  labs(title = "Title", fill = y, x= x, y= y) +
  theme_scg()
}
}
