% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_display.R
\name{colour_display}
\alias{colour_display}
\title{Diplay Colour Options}
\usage{
colour_display(pal_name, n, assign)
}
\arguments{
\item{pal_name}{Name of pal_name to visualise}

\item{n}{Number of desired colours.}

\item{assign}{Vector of corresponding levels/cetgories}
}
\value{
a ggplot2 graph
}
\description{
Visualise colour options
}
\examples{
# View full pallette
colour_display("polUK")

# View individual colour
colour_display("Jaffa")

# View all individual colours
colour_display("All")

# View sequential colour pal_name with 7 levels
colour_display("seqGreen", 7)

# View diverging colour pal_name with 5 levels with assigned values
colour_display("divBlueGreen", 5, c("Very Likely","Likely","Neutral","Unlikely","Very Unlikely"))

}
