% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{compile}
\alias{compile}
\title{Compile Multiple Data Frames in One}
\usage{
compile(
  data,
  row_vars,
  col_vars,
  weight,
  format = c("csv", "statistics"),
  save = TRUE,
  name = "table"
)
}
\arguments{
\item{data}{A data frame containing survey data. This parameter is required.}

\item{row_vars}{Independent variable represented in rows (on the side). This parameter is required.}

\item{col_vars}{Dependent variable represented in columns (at the top). This parameter is required.}

\item{weight}{Variable containing weight factors. This variable is optional.}

\item{format}{Formatting options to return either a csv or statistics data frame (default = \code{"csv"}).}

\item{save}{Logical, if \code{TRUE}, df will be saved to .csv file (default = \code{TRUE}).}

\item{name}{Name of df to be saved to project directory (default = \code{"table"}).}
}
\value{
.csv file
}
\description{
Produce a csv containing a full set of crosstabs or statistics for each 2x2 variables.
}
\examples{
\dontrun{
# Create crosstabs with csv output using weighted data.
compile(dataset,
       row_vars = c("Q1", "Q2"),
       col_vars = c("Gender", "VI"),
       weight = "wgtvar",
       name = "crosstabs")
}
}
