% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab.R
\name{crosstab}
\alias{crosstab}
\title{Crosstab / Contingency Tables}
\usage{
crosstab(
  data,
  row_var,
  col_var,
  weight,
  totals = TRUE,
  round_decimals = NULL,
  statistics = TRUE,
  plot = TRUE,
  format = c("df_long", "df_wide", "csv", "statistics"),
  convert_to = c("percent", "frequency")
)
}
\arguments{
\item{data}{A data frame containing survey data. This parameter is required.}

\item{row_var}{Independent variable represented in rows (on the side). This parameter is required.}

\item{col_var}{Dependent variable represented in columns (at the top). This parameter is required.}

\item{weight}{Variable containing weight factors. This variable is optional.}

\item{totals}{Logical, if totals is \code{TRUE}, df with totals column is included (default = \code{TRUE}).}

\item{round_decimals}{Set the number of decimal points for the data (default = \code{NULL}).}

\item{statistics}{Logical, if statistics is \code{TRUE}, Chi-Square, DF, Cramer's V, and p-value are printed (default = \code{TRUE}).}

\item{plot}{Logical, if plot is \code{TRUE}, a chart is available to be viewed (default = \code{TRUE}).}

\item{format}{Formatting options to return either a long or wide data frame (default = \code{"df_long"}).}

\item{convert_to}{Conversion options to return either a percentages or frequencies (default = \code{"percent"}).}
}
\value{
Crosstabs held in a data frame containing row-wise percentages (\%) and col-wise totals (n)
}
\description{
Produce crosstabs/contingency tables to view individually and perform further analysis on.
}
\examples{
\dontrun{
# Create crosstabs with long df output using weighted data.
df <- crosstab(dataset,
              row_var = "Q1",
              col_var = "Gender",
              weight = "wgtvar",
              totals = FALSE,
              round_decimal=2)
}
}
