% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab.R
\name{crosstab}
\alias{crosstab}
\title{Crosstab / Contingency Tables}
\usage{
crosstab(
  data,
  rowVar,
  colVar,
  weight = NULL,
  totals = TRUE,
  round_decimals = NULL,
  statistics = FALSE,
  plot = FALSE,
  format = c("df_long", "df_wide", "csv", "statistics"),
  convert_to = c("percent", "frequency"),
  yLab = "Population (\%)",
  adjustX = c("no", "yes")
)
}
\arguments{
\item{data}{A data frame containing survey data. This parameter is required.}

\item{rowVar}{Independent variable represented in rows (on the side). This parameter is required.}

\item{colVar}{Dependent variable represented in columns (at the top). This parameter is required.}

\item{weight}{Variable containing weight factors. This variable is optional.}

\item{totals}{Logical, if totals is \code{TRUE}, df with totals column is included (default = \code{TRUE}).}

\item{round_decimals}{Set the number of decimal points for the data (default = \code{NULL}).}

\item{statistics}{Logical, if statistics is \code{TRUE}, Chi-Square, DF, Cramer's V, and p-value are printed (default = \code{TRUE}).}

\item{plot}{Logical, if plot is \code{TRUE}, a chart is available to be viewed (default = \code{TRUE}).}

\item{format}{Formatting options to return either a long or wide data frame (default = \code{"df_long"}).}

\item{convert_to}{Conversion options to return either a percentages or frequencies (default = \code{"percent"}).}

\item{yLab}{Y axis title. Default = "Population (\%)".}

\item{adjustX}{If "yes", the values on the x-axis will turn 45 degress. The default is "no" which keeps values horiztonal.}
}
\value{
Crosstabs held in a data frame containing row-wise percentages (\%) and col-wise totals (n)
}
\description{
Produce crosstabs/contingency tables to view individually and perform further analysis on.
}
\examples{
\dontrun{
# Create crosstabs with long df output using weighted data.
df <- crosstab(dataset,
              rowVar = "Q1",
              colVar = "Gender",
              weight = "wgtvar",
              totals = FALSE,
              round_decimal=2)
}
}
