% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_vars.R
\name{grid_vars}
\alias{grid_vars}
\title{Pivot Binary Grid Survey Questions Longer}
\usage{
grid_vars(data, vars, group = NULL, weight = NULL)
}
\arguments{
\item{data}{A data frame containing survey data.}

\item{vars}{A list mapping original column names to new variable names. This is
required and allows multiple variables. For example, \code{list(Q1a = "Art", Q1b = "Automobiles")}.}

\item{group}{An optional variable for comparing across groups. Only one group can
be specified.}

\item{weight}{An optional variable containing weight factors for the analysis.}
}
\value{
A data frame in a long format, with each row representing a response to a
binary grid question. The data frame includes frequencies and optionally percentages
if groups and/or weights are specified.
}
\description{
The \code{grid_vars} function transforms binary grid survey questions into a longer format,
enabling comparisons across different variables or groups. It supports optional grouping
and weighting.
}
\details{
The function first converts specified binary grid questions into a long format using
\code{tidyr::pivot_longer}. It then uses \code{grp_freq} to calculate frequencies and, if
applicable, percentages for each question-response pair. This allows for an in-depth
analysis of binary grid survey questions, especially when combined with group and weight parameters.
}
\examples{
\dontrun{
  # Example: Convert grid questions to long format and analyze by gender
  vars <- list(Q1a = "Art", Q1b = "Automobiles", Q1c = "Birdwatching")
  df <- grid_vars(dataset, vars = vars, group = "gender", weight = "wgtvar")
}

}
