% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_freq.R
\name{grp_freq}
\alias{grp_freq}
\title{Grouped Frequencies}
\usage{
grp_freq(
  data,
  groups,
  weight = NULL,
  set_names = NULL,
  addPercent = c("no", "yes"),
  groupsPercent = NULL,
  round_decimals = NULL
)
}
\arguments{
\item{data}{A data frame containing survey data. This parameter is required.}

\item{groups}{Vector of variables being grouped. This parameter is required and permits one or many variables.}

\item{weight}{Variable containing weight factors. This variable is optional.}

\item{set_names}{Vector of column names. This paramenter is optional.}

\item{addPercent}{Get percent of frequency (Options: "no" or "yes"). This parameter is optional, defaulted to "No", and permits one or many variables.}

\item{groupsPercent}{Vector of variables for percent of frequency. This parameter is optional and permits none (for total), one or many variables.}

\item{round_decimals}{Numeric value to round numeric data by x number of decimals places. Default does not round. This parameter is optional.}
}
\value{
A data frame containing frequencies by group
}
\description{
Calculates the frequency of a variable by groups in survey data.
}
\examples{
\dontrun{
# Sum unweighted data by group
df <- grp_freq(dataset,
               group = c("age_categories","gender"))
# Sum weighted data by group and change the column names
df <- grp_freq(dataset,
               group = c("age_categories","gender"),
               weight = "wgtvar",
               set_names = c("Age","Gender","n"))
}
}
