% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_freq.R
\name{grp_freq}
\alias{grp_freq}
\title{Grouped Frequencies}
\usage{
grp_freq(
  data,
  groups,
  weight,
  set_names,
  addPercent = c("no", "yes"),
  groupsPercent,
  round_decimals = NULL
)
}
\arguments{
\item{data}{A data frame containing survey data. This parameter is required.}

\item{groups}{Vector of variables being grouped. This parameter is required and permits one or many variables.}

\item{weight}{Variable containing weight factors. This variable is optional.}

\item{set_names}{Vector of column names. This paramenter is optional.}

\item{addPercent}{Get percent of frequency (Options: "no" or "yes"). This parameter is optional, defaulted to "No", and permits one or many variables.}

\item{groupsPercent}{Vector of variables for percent of frequency. This parameter is optional and permits none (for total), one or many variables.}

\item{round_decimals}{Numeric value to round numeric data by x number of decimals places. Default does not round. This parameter is optional.}
}
\value{
A data frame containing frequencies by group
}
\description{
Calculates the frequency of a variable by groups in survey data.
}
\examples{
\dontrun{
# Sum unweighted data by group
df <- grp_freq(dataset,
               group = c("age_categories","gender"))

#    age_categories gender Freq
# 1        18 to 24   Male   84
# 2        25 to 34   Male  155
# 3        35 to 44   Male  180
# 4        45 to 54   Male  115
# 5        55 to 54   Male   38
# 6             65+   Male    6
# 7        18 to 24 Female   83
# 8        25 to 34 Female  153
# 9        35 to 44 Female  129
# 10       45 to 54 Female   57
# 11       55 to 54 Female   24
# 12            65+ Female    6

# Sum weighted data by group and change the column names
df <- grp_freq(dataset,
               group = c("age_categories","gender"),
               weight = "wgtvar",
               set_names = c("Age","Gender","n"))

#         Age Gender         n
# 1  18 to 24   Male  43.40468
# 2  25 to 34   Male 115.36419
# 3  35 to 44   Male 160.55258
# 4  45 to 54   Male 187.29274
# 5  55 to 54   Male  61.88803
# 6       65+   Male  30.90000
# 7  18 to 24 Female  42.88796
# 8  25 to 34 Female 109.82791
# 9  35 to 44 Female 115.06268
# 10 45 to 54 Female  92.83205
# 11 55 to 54 Female  39.08718
# 12      65+ Female  30.90000
}
}
