% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bigfive.R
\name{plot_bigfive}
\alias{plot_bigfive}
\title{Big Five Personality Radar Plots}
\usage{
plot_bigfive(data, bigfive, group = NULL, weight = NULL)
}
\arguments{
\item{data}{A data frame containing survey data. This parameter is required.}

\item{bigfive}{A vector with the column names representing the Big Five personality trait scores. This parameter is required.}

\item{group}{A variable overlay to compare between groups. This parameter is optional.}

\item{weight}{Variable containing weight factors. This variable is optional.}
}
\value{
A radar plot created using ggplot2, displaying the average scores of the Big Five personality traits.
}
\description{
Visualises the average scores of the Big Five personality traits on a radar graph.
}
\examples{
\dontrun{
# Create a radar plot using age groups
plot_bigfive(dataset,
             bigfive = c("Neuroticism", "Extroversion", "Openness",
                          "Agreeableness", "Conscientiousness"),
             group = "age_categories",
             weight = "wgtvar")
}
}
