% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bigfive.R
\name{plot_bigfive}
\alias{plot_bigfive}
\title{Big Five Personality Radar Plots}
\usage{
plot_bigfive(data, big_five, group, weight)
}
\arguments{
\item{data}{A data frame containing survey data. This parameter is required.}

\item{big_five}{A vector with the columns containing the big five scores. This parameter is required.}

\item{group}{A variable overlay to compare between groups. This parameter is optional.}

\item{weight}{Variable containing weight factors. This variable is optional.}
}
\value{
ggplot2 plot of the big five scores
}
\description{
Visualise the personality trait scores on a radar graph.
}
\examples{
\dontrun{
# Create plot using age groups
plot_bigfive(dataset,
             big_five = c("Neuroticism", "Extroversion", "Openness",
                          "Agreeableness", "Conscientiousness"),
             group = "age_categories",
             weight = "wgtvar")
}
}
