% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_popn.R
\name{plot_popn}
\alias{plot_popn}
\title{Population Profile Plot}
\usage{
plot_popn(data, age_groups, gender, group, weight, age_int)
}
\arguments{
\item{data}{A data frame containing survey data. This parameter is required.}

\item{age_groups}{Age group variable. This parameter is required.}

\item{gender}{Gender variable. This parameter is required.}

\item{group}{A variable overlay to compare between groups. This parameter is optional.}

\item{weight}{Variable containing weight factors. This variable is optional.}

\item{age_int}{Actual age vraiable (numeric) to view average age.}
}
\value{
ggplot2 plot of the population structure
}
\description{
Visualise the population profile of survey data using age categories and gender.
}
\examples{
\dontrun{
# Create plot using age groups and age intervals included to return average age
plot_popn(dataset,
          age_group = "age_categories",
          gender = "gender",
          weight = "wgtvar",
          age_int = "age")
}
}
