% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sankey.R
\name{plot_sankey}
\alias{plot_sankey}
\title{Visualise Data Flow with a Sankey Diagram}
\usage{
plot_sankey(
  data,
  source,
  target,
  value,
  units = "",
  colours = colour_pal("catExtended"),
  fontSize = 20,
  fontFamily = "Calibri",
  nodeWidth = 20,
  nodePadding = 10,
  margin = list(left = 0, right = 0),
  width = 1200,
  height = 800,
  shiftLabel = NULL
)
}
\arguments{
\item{data}{A data frame containing the flow data, where each row represents a unique flow
from a source to a target node.}

\item{source}{The column in \code{data} representing the source nodes.}

\item{target}{The column in \code{data} representing the target nodes.}

\item{value}{The column in \code{data} representing the flow's magnitude or volume.}

\item{units}{A string indicating the units for \code{value}, displayed in the tooltip.
Defaults to an empty string, implying no units.}

\item{colours}{A list or vector of colours for nodes and links in the diagram, either
as hexadecimal codes or colour names. Can be a named list for specific node colour mapping.}

\item{fontSize}{Font size for text in the diagram.}

\item{fontFamily}{Font family for text in the diagram.}

\item{nodeWidth}{Width of nodes in the diagram.}

\item{nodePadding}{Padding between nodes in the diagram.}

\item{margin}{Margins around the diagram, specified as a list.}

\item{width}{Width of the plot in pixels.}

\item{height}{Height of the plot in pixels.}

\item{shiftLabel}{Numeric value to adjust the position of labels; \code{NA} for auto-placement.}
}
\value{
An interactive Sankey diagram as an HTML widget, which can be used in R Markdown documents,
Shiny applications, or the R console.
}
\description{
\code{plot_sankey} enhances the \href{https://www.rdocumentation.org/packages/networkD3/versions/0.4/topics/sankeyNetwork}{\code{networkD3::sankeyNetwork}}
function, utilising \href{https://www.htmlwidgets.org/}{\code{htmlwidgets}} to create an interactive Sankey diagram,
which is a specific type of flow diagram. It's especially effective for illustrating
data movement or transfer between different entities (nodes), such as the distribution
of votes flowing between parties across elections or in preferential voting systems.
}
\examples{
\dontrun{
  # Example: Visualizing the flow of votes between parties
  plot_sankey(data = election_data,
              source = "Party_Previous_Election",
              target = "Party_Current_Election",
              value = "Number_of_Votes",
              colours = c("Labour" = "red", "Green" = "green", "National" = "blue"))
}

}
