% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sankey.R
\name{plot_sankey}
\alias{plot_sankey}
\title{Sankey Diagram Visualisation}
\usage{
plot_sankey(
  data,
  source,
  target,
  value,
  units = "",
  colours = colour_pal("catExtended"),
  fontSize = 20,
  fontFamily = "Calibri",
  nodeWidth = 20,
  nodePadding = 10,
  margin = list(left = 0, right = 0),
  width = 1200,
  height = 800,
  shiftLabel = NA
)
}
\arguments{
\item{data}{A data frame containing the flow data, where each row represents a unique flow from a source to a target.}

\item{source}{The column in the data frame representing the source node of the flow.}

\item{target}{The column in the data frame representing the target node of the flow.}

\item{value}{The column in the data frame representing the magnitude or volume of the flow between the source and target.}

\item{units}{A string indicating the units of the 'value'. This is displayed in the tooltip of the diagram.
The default is an empty string, indicating that no units are specified.}

\item{colours}{A list or vector of colours (in hexadecimal format or colour names) to be used for the nodes and links
in the diagram. If a named list is provided, the names should correspond to the node names for specific
colour mapping.}

\item{fontSize}{The font size for text in the diagram. Default is set to 20.}

\item{fontFamily}{The font family for text in the diagram. Default is "Calibri".}

\item{nodeWidth}{The width of the nodes (entities) in the diagram. Default is 20.}

\item{nodePadding}{The padding between nodes in the diagram. Default is 10.}

\item{margin}{A list specifying the margins around the diagram. Default is \code{list("left" = 0, "right" = 0)}.
Increase the margin if the node names are long and go off the page.}

\item{width}{The width of the plot in pixels. Default is 1200.}

\item{height}{The height of the plot in pixels. Default is 800.}

\item{shiftLabel}{A numeric value controlling the positioning of labels on the nodes. When set to \code{NA},
labels are automatically placed outside the nodes. If the automatic placement results in uneven label
distribution, this parameter can be adjusted to achieve the desired layout.}
}
\value{
A Sankey diagram visualising the flow of data as specified by the parameters. The diagram is interactive,
allowing for the exploration of individual flows and nodes.
}
\description{
\code{plot_sankey} enhances
\href{https://www.rdocumentation.org/packages/networkD3/versions/0.4/topics/sankeyNetwork}{\code{networkD3::sankeyNetwork}}
and \href{https://www.htmlwidgets.org/}{\code{htmlwidgets}} to create a Sankey diagram, a specific type of flow diagram
that visually represents the movement or flow of data between different entities (referred to as "nodes").
It is particularly useful for illustrating the distribution of votes flowing between parties across elections
or in preferential voting systems.
}
\examples{
\dontrun{
# Prepare data
split <- grp_freq(df,
                  c("Q1a","Q1c"),
                  "wgtvar",
                  c("Party_Vote","Candidate_Vote","Votes")) \%>\%
         mutate(Votes = Votes/sum(Votes))
#           Party_Vote    Candidate_Vote        Votes
# 1       Labour Party      Labour Party 0.1036125880
# 2     National Party      Labour Party 0.1040036893
# 3        Green Party      Labour Party 0.0320547032
# ...

# Create sankey
plot_sankey(data = split,
            source = "Party_Vote",
            target = "Candidate_Vote",
            value = "Votes",
            colours = colour_prep(df, c("Q1a", "Q1c"), pal_name = "polNZ"))
}
}
