% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sankey.R
\name{plot_sankey}
\alias{plot_sankey}
\title{Sankey Plot}
\usage{
plot_sankey(
  data,
  source,
  target,
  value,
  units = "",
  colours,
  fontSize = 20,
  fontFamily = "Calibri",
  nodeWidth = 20,
  nodePadding = 10,
  margin = list(left = 0, right = 0),
  width = 1200,
  height = 800,
  shiftLabel = NA
)
}
\arguments{
\item{data}{A data frame containing flow data and percentages. This parameter is required.}

\item{source}{From variable (left side of sankey). This parameter is required.}

\item{target}{To variable (right side of sankey). This parameter is required.}

\item{value}{Percentage/Value flowing between source and target. This parameter is required.}

\item{units}{Unit of value to display in tooltip (requires single quotation marks around double
quotations). Default = '""'}

\item{colours}{List of colours in d3.js format (requires single quotation marks around internal hex codes)
double quotation marks. This variable is required.}

\item{fontSize}{Default = 24}

\item{fontFamily}{Default = "Calabri"}

\item{nodeWidth}{Default = 20}

\item{nodePadding}{Default = 10}

\item{margin}{Default = list("left"=0, "right"=400)}

\item{width}{Default = 1200}

\item{height}{Default = 800}

\item{shiftLabel}{the default will automatically place labels to outside of the sankey, however if there
are unequal numbers of labels, enter in a numeric value to control which labels shift from inside to out.}
}
\value{
sankey plot
}
\description{
Visualise the flow of data with a sankey.
}
\examples{
\dontrun{
# Prepare data
split <- grp_freq(df,
                  c("Q1a","Q1c"),
                  "wgtvar",
                  c("Party_Vote","Candidate_Vote","Votes")) \%>\%
         mutate(Votes = Votes/sum(Votes))
#           Party_Vote    Candidate_Vote        Votes
# 1       Labour Party      Labour Party 0.1036125880
# 2     National Party      Labour Party 0.1040036893
# 3        Green Party      Labour Party 0.0320547032
# ...

# Create sankey
plot_sankey(data = split,
            source = "Party_Vote",
            target = "Candidate_Vote",
            value = "Votes",
            colours = '"#d62f24","#378cc5","#4fbe57","#fee043","#9b0d0c",
                       "#3e3e3f","#e69138ff","#b7b7b7ff","#cdcdd1","#cdcdd1"')
}
}
