% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_levels.R
\name{remove_levels}
\alias{remove_levels}
\title{Remove Unused Levels from Factors}
\usage{
remove_levels(data)
}
\arguments{
\item{data}{A data frame containing survey data. This parameter is required.}
}
\value{
data frame
}
\description{
Remove unused levels from factors across a complete data frame. NB non-factor columns will be moved
to the beginning of the data frame.
}
\examples{
\dontrun{
df <- remove_levels(df)
}
}
